Release 11.02.26
git-svn-id: http://svn.code.sf.net/p/netpbm/code/stable@5277 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/pbm/mdatopbm.c
@@ -241,8 +241,8 @@ main(int argc, char **argv) {
 
     {
         short yy;
-        pm_readlittleshort(infile, &yy); nInRows = yy;
-        pm_readlittleshort(infile, &yy); nInCols = yy;
+        pm_readlittleshort(infile, &yy); nInRows = (unsigned short)yy;
+        pm_readlittleshort(infile, &yy); nInCols = (unsigned short)yy;
     }
     
     nOutCols = 8 * nInCols;
